move fetch function to cmd_funcs
